mark tests that use storage
