+++ tests/api/test_files.py
@@ -22,6 +22,8 @@ def draft(db, default_account):
     }
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 def test_file_filtering(api_client, uploaded_file_ids, draft):
     # Attach the files to a draft and search there
     draft["file_ids"] = uploaded_file_ids
@@ -62,6 +64,8 @@ def test_file_filtering(api_client, uploaded_file_ids, draft):
     assert len(results) == 2
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 def test_attachment_has_same_id(api_client, uploaded_file_ids, draft):
     attachment_id = uploaded_file_ids.pop()
     draft["file_ids"] = [attachment_id]
@@ -71,6 +75,8 @@ def test_attachment_has_same_id(api_client, uploaded_file_ids, draft):
     assert attachment_id in [x["id"] for x in draft_resp["files"]]
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 def test_delete(api_client, uploaded_file_ids, draft):
     non_attachment_id = uploaded_file_ids.pop()
     attachment_id = uploaded_file_ids.pop()
@@ -93,6 +99,8 @@ def test_delete(api_client, uploaded_file_ids, draft):
     assert data["id"] == attachment_id
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 @pytest.mark.parametrize("filename", FILENAMES)
 def test_get_with_id(api_client, uploaded_file_ids, filename):
     # See comment in uploaded_file_ids()
@@ -124,6 +132,8 @@ def test_get_invalid(api_client, uploaded_file_ids):
     assert r.status_code == 400
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 @pytest.mark.parametrize("filename", FILENAMES)
 def test_download(api_client, uploaded_file_ids, filename):
     # See comment in uploaded_file_ids()
