+++ tests/imap/test_folder_sync.py
@@ -50,6 +50,8 @@ def trash_folder(db, default_account):
     )
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 def test_initial_sync(db, generic_account, inbox_folder, mock_imapclient):
     # We should really be using hypothesis.given() to generate lots of
     # different uid sets, but it's not trivial to ensure that no state is
@@ -77,6 +79,8 @@ def test_initial_sync(db, generic_account, inbox_folder, mock_imapclient):
     }
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 def test_new_uids_synced_when_polling(
     db, generic_account, inbox_folder, mock_imapclient
 ):
@@ -264,6 +268,8 @@ def raise_imap_error(self):
     raise IMAP4.error("Unexpected IDLE response")
 
 
+@pytest.mark.usefixtures("blockstore_backend")
+@pytest.mark.parametrize("blockstore_backend", ["disk", "s3"], indirect=True)
 def test_gmail_initial_sync(db, default_account, all_mail_folder, mock_imapclient):
     uid_dict = uids.example()
     mock_imapclient.add_folder_data(all_mail_folder.name, uid_dict)
